+++ src/main/java/ca/wescook/homesteadcompanion/nutrition/NutritionStorage.java
@@ -14,16 +14,16 @@ public class NutritionStorage implements Capability.IStorage<INutrition> {
 	public NBTBase writeNBT(Capability<INutrition> capability, INutrition instance, EnumFacing side) {
 		NBTTagCompound playerData = new NBTTagCompound();
 		for (Nutrient nutrient : NutrientList.get())
-			playerData.setInteger(nutrient.name, instance.get(nutrient));
+			playerData.setFloat(nutrient.name, instance.get(nutrient));
 		return playerData;
 	}
 
 	// Load serialized data from disk
 	@Override
 	public void readNBT(Capability<INutrition> capability, INutrition instance, EnumFacing side, NBTBase nbt) {
-		HashMap<Nutrient, Integer> clientNutrients = new HashMap<Nutrient, Integer>(); // Create new map
+		HashMap<Nutrient, Float> clientNutrients = new HashMap<Nutrient, Float>(); // Create new map
 		for (Nutrient nutrient : NutrientList.get()) {
-			Integer value = ((NBTTagCompound) nbt).getInteger(nutrient.name); // Read ints from packet
+			Float value = ((NBTTagCompound) nbt).getFloat(nutrient.name); // Read values from packet
 			clientNutrients.put(nutrient, value); // Add to map
 		}
 		instance.set(clientNutrients); // Replace nutrient data wih map
